requirements.txt
Django
django-rest-framework
django-adminlte2
crispy-bootstrap5
pytz

